Removed trigger datastream

+++ Derivatives.php
@@ -225,27 +225,29 @@ class Derivative {
   }
 
   function Scholar_PDFA($dsid = 'PDF', $label = 'PDF') {
-    $this->log->lwrite('Starting processing', 'PROCESS_DATASTREAM', $this->pid, $dsid);
-    try {
-      $output_file = $this->temp_file . '_Scholar_PDFA.xml';
-      if ($this->mimetype == 'application/pdf') {
-        exec("gs -dPDFA -dBATCH -dNOPAUSE -sDEVICE=pdfwrite -sOutputFile=$output_file $this->temp_file", $pdfa_output, $return);
+    if ($this->incoming_dsid == 'OBJ') {
+      $this->log->lwrite('Starting processing', 'PROCESS_DATASTREAM', $this->pid, $dsid);
+      try {
+        $output_file = $this->temp_file . '_Scholar_PDFA.xml';
+        if ($this->mimetype == 'application/pdf') {
+          exec("gs -dPDFA -dBATCH -dNOPAUSE -sDEVICE=pdfwrite -sOutputFile=$output_file $this->temp_file", $pdfa_output, $return);
+        }
+        else {
+          exec("java -jar /opt/jodconverter-core-3.0-beta-4/lib/jodconverter-core-3.0-beta-4.jar $this->temp_file $output_file", $pdfa_output, $return);
+        }
+        $pdfa_datastream = new NewFedoraDatastream($dsid, 'M', $this->object, $this->fedora_object->repository);
+        $pdfa_datastream->setContentFromFile($output_file);
+        $pdfa_datastream->label = $label;
+        $pdfa_datastream->mimetype = 'application/pdf';
+        $pdfa_datastream->state = 'A';
+        $this->object->ingestDatastream($pdfa_datastream);
+        unlink($output_file);
+      } catch (Exception $e) {
+        $this->log->lwrite("Could not create the $dsid derivative!", 'FAIL_DATASTREAM', $this->pid, $dsid, NULL, 'ERROR');
+        unlink($output_file);
       }
-      else {
-        exec("java -jar /opt/jodconverter-core-3.0-beta-4/lib/jodconverter-core-3.0-beta-4.jar $this->temp_file $output_file", $pdfa_output, $return);
-      }
-      $pdfa_datastream = new NewFedoraDatastream($dsid, 'M', $this->object, $this->fedora_object->repository);
-      $pdfa_datastream->setContentFromFile($output_file);
-      $pdfa_datastream->label = $label;
-      $pdfa_datastream->mimetype = 'application/pdf';
-      $pdfa_datastream->state = 'A';
-      $this->object->ingestDatastream($pdfa_datastream);
-      unlink($output_file);
-    } catch (Exception $e) {
-      $this->log->lwrite("Could not create the $dsid derivative!", 'FAIL_DATASTREAM', $this->pid, $dsid, NULL, 'ERROR');
-      unlink($output_file);
+      return $return;
     }
-    return $return;
   }
 
 }
